pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asthma-map"
version = "0.1.0"
description = "NYC Childhood Asthma Prevention Failure Map"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.0.0",
    "geopandas>=0.14.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "shapely>=2.0.0",
    "matplotlib>=3.8.0",
    "censusgeocode>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]

